+++ core/auth/email.go
@@ -107,19 +107,19 @@ func (a *emailAuthImpl) signUp(authType model.AuthType, appType model.Applicatio
 	return "verification code sent successfully", emailCredValueMap, nil
 }
 
-func (a *emailAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, *bool, error) {
+func (a *emailAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, error) {
 	verifyEmail := a.getVerifyEmail(accountAuthType.AuthType)
 	if verifyEmail {
 		//check is verified
 		if !accountAuthType.Credential.Verified {
-			return "", nil, errors.ErrorAction("not verified", "", nil)
+			return "", errors.ErrorAction("not verified", "", nil).SetStatus(utils.ErrorStatusUnverified)
 		}
 	}
 
 	//get stored credential
 	storedCreds, err := mapToEmailCreds(accountAuthType.Credential.Value)
 	if err != nil {
-		return "", nil, errors.WrapErrorAction("error on map to email creds", "", nil, err)
+		return "", errors.WrapErrorAction("error on map to email creds", "", nil, err)
 	}
 
 	//get request credential
@@ -129,18 +129,17 @@ func (a *emailAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, 
 	var sPasswordParams signInPasswordCred
 	err = json.Unmarshal([]byte(creds), &sPasswordParams)
 	if err != nil {
-		return "", nil, errors.WrapErrorAction("error getting sign_in password creds", "", nil, err)
+		return "", errors.WrapErrorAction("error getting sign_in password creds", "", nil, err)
 	}
 	requestPassword := sPasswordParams.Password
 
 	//compare stored and requets ones
 	err = bcrypt.CompareHashAndPassword([]byte(storedCreds.Password), []byte(requestPassword))
 	if err != nil {
-		return "", nil, errors.WrapErrorAction("bad credentials", "", nil, err)
+		return "", errors.WrapErrorAction("bad credentials", "", nil, err).SetStatus(utils.ErrorStatusInvalid)
 	}
 
-	valid := true
-	return "", &valid, nil
+	return "", nil
 }
 
 func (a *emailAuthImpl) getVerifyEmail(authType model.AuthType) bool {