+++ core/auth/username.go
@@ -30,8 +30,8 @@ func (a *usernameAuthImpl) verify(credential *model.Credential, verification str
 	return nil, errors.New(logutils.Unimplemented)
 }
 
-func (a *usernameAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, *bool, error) {
-	return "", nil, nil
+func (a *usernameAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, error) {
+	return "", nil
 }
 func (a *usernameAuthImpl) sendVerify(authType model.AuthType, identifier string, credential *model.Credential, verifyWaitTime int64, l *logs.Log) (map[string]interface{}, error) {
 	return nil, nil