+++ driven/storage/model_auth.go
@@ -22,10 +22,10 @@ type loginSession struct {
 
 	DeviceID string `bson:"device_id"`
 
-	IPAddress    string                 `bson:"ip_address"`
-	AccessToken  string                 `bson:"access_token"`
-	RefreshToken string                 `bson:"refresh_token"`
-	Params       map[string]interface{} `bson:"params"`
+	IPAddress     string                 `bson:"ip_address"`
+	AccessToken   string                 `bson:"access_token"`
+	RefreshTokens []string               `bson:"refresh_tokens"`
+	Params        map[string]interface{} `bson:"params"`
 
 	Expires time.Time `bson:"expires"`